Create testing data collection
Here we are asking the users for their input that if they are tested positive or not for covid and then we are adding/saving the data in the csv files. We have to make sure that the user input is valid otherwise the program will prompt the user to enter the valid input. Further, we will test the quality of the data using a condition that if the quality is greater than 0.95, it is accepted. Lastly, we will test the calibration time and that should be less than 5.